pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "unicorns"
version = "0.1.0"
authors = [
    { name="Tsz Kiu Pang", email="osamupang@gmail.com" },
]
description = "The Last Unicorns, for solo piano"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "abjad",
    "abjad-ext-nauert",
    "pang",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
test = [
    "black",
    "flake8",
    "isort",
    "pytest",
]
